long time since last commit, sorry, here's what's been added: progress tracking on the lessonlist screen works, though it's not 100% consistent--still figuring out why. progress tracking on play screen works

+++ screens/StudySetScreen.js
@@ -15,7 +15,10 @@ function StudySetScreen(props) {
     const [isFirstLaunch, setIsFirstLaunch] = useState(false);
 
     async function checkFirstLaunch() {
-        AsyncStorage.clear()
+      /*   const asyncStorageKeys = await AsyncStorage.getAllKeys();
+        if (asyncStorageKeys.length > 0) {
+            AsyncStorage.clear();
+        ***REMOVED*** */
         try {
             await AsyncStorage
                 .getItem('alreadyLaunched')
@@ -32,13 +35,22 @@ function StudySetScreen(props) {
     ***REMOVED***
 
     function setProgress() {
-        var lesson;
+        //old
+        /* var lesson;
         for (i = 0; i < STUDYSETS.length; i++) {
             for (j = 0; j < STUDYSETS[i].lessonList.length; j++) {
                 lesson = STUDYSETS[i].lessonList[j].id
                 setAsyncValue(lesson, 'incomplete');
             ***REMOVED***
-        ***REMOVED*** 
+        ***REMOVED***  */
+
+        var progress = {***REMOVED***;
+        for (i = 0; i < STUDYSETS.length; i++) {
+            for (j = 0; j < STUDYSETS[i].lessonList.length; j++) {
+                progress[STUDYSETS[i].lessonList[j].id] = 'incomplete'
+            ***REMOVED***
+        ***REMOVED***
+        setAsyncValue("progress", JSON.stringify(progress));
     ***REMOVED***
 
     async function setAsyncValue(key, mark) {
@@ -53,6 +65,7 @@ function StudySetScreen(props) {
     ***REMOVED***
 
     useEffect(() => {
+        console.log("study set screen use effect")
         checkFirstLaunch();
     ***REMOVED***, [])
